Added ServiceTransportTest.java; migrated test_remote_node_died onto it
